Extract common flink functions
